requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
paramiko==2.11.0
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
